Moved growup to it's own class
