+++ ValheimPlus/ValheimPlus.csproj
@@ -209,6 +209,8 @@
     <Compile Include="GameClasses\Character.cs" />
     <Compile Include="GameClasses\CharacterDrop.cs" />
     <Compile Include="GameClasses\Demister.cs" />
+    <Compile Include="GameClasses\Egg.cs" />
+    <Compile Include="GameClasses\Growup.cs" />
     <Compile Include="GameClasses\InventoryGrid.cs" />
     <Compile Include="GameClasses\LuredWisp.cs" />
     <Compile Include="GameClasses\PickableItem.cs" />
